hrms-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
pydantic==2.5.0
pydantic-settings==2.1.0
redis==5.0.1
celery==5.3.4
stripe==7.8.0
sendgrid==6.11.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
opencv-python==4.8.1.78
face-recognition==1.3.0
numpy==1.24.4
pillow==10.1.0
boto3==1.34.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2